// src/enums.ts
export enum ProgramToken {
  DANK = "dank_mint",
  CFX = "cfx_mint",
}

export enum ProgramAuction {
  CFXUSDX = "cfx_usdx_da",
  USDXCFX = "usdx_cfx_da",
  DANKCFX = "dank_cfx_da",
  USDXDANK = "usdx_dank_da",
}

export enum ProgramOracleManager {
  USDXUSD = "usdx_usd_oracle_manager",
  FXUSD = "fx_usd_oracle_manager",
  SOLUSD = "sol_usd_oracle_manager",
}